requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-APScheduler
Flask-WTF
email-validator
python-dotenv